requirements.txt
Flask
byaldi
torch
torchvision
google-generativeai
openai
docx2pdf
qwen-vl-utils
vllm>=0.6.1.post1; sys_platform != 'darwin'
mistral_common>=1.4.1
einops
mistral_common[opencv]
mistral_common
mistral_inference
groq
markdown
hf_transfer
ollama
streamlit
langchain
chromadb
unstructured[all-docs]
pytesseract
Pillow
python-dotenv